+++ Source/ProjectFirma.Web/ProjectFirma.Web.csproj
@@ -389,8 +389,8 @@
     <Compile Include="Views\ProjectCreate\Watershed.cs" />
     <Compile Include="Views\ProjectCreate\WatershedViewData.cs" />
     <Compile Include="Views\ProjectCreate\WatershedViewModel.cs" />
-    <Compile Include="Views\Project\FundingRequestSheet.cs" />
-    <Compile Include="Views\Project\FundingRequestSheetViewData.cs" />
+    <Compile Include="Views\Project\ForwardLookingFactSheet.cs" />
+    <Compile Include="Views\Project\ForwardLookingFactSheetViewData.cs" />
     <Compile Include="Views\Project\FundingSourceRequestAmount.cs" />
     <Compile Include="Views\Shared\GoogleChartAnnotations.cs" />
     <Compile Include="Views\Shared\GoogleChartAxis.cs" />
@@ -1126,8 +1126,8 @@
     <Compile Include="Views\Project\EditFeaturedProjectsViewModel.cs" />
     <Compile Include="Views\Project\PerformanceMeasureSubcategoriesCalendarYearReportedValue.cs" />
     <Compile Include="Views\Project\PerformanceMeasureSubcategoriesTotalReportedValue.cs" />
-    <Compile Include="Views\Project\FactSheet.cs" />
-    <Compile Include="Views\Project\FactSheetViewData.cs" />
+    <Compile Include="Views\Project\BackwardLookingFactSheet.cs" />
+    <Compile Include="Views\Project\BackwardLookingFactSheetViewData.cs" />
     <Compile Include="Views\Project\FeaturedList.cs" />
     <Compile Include="Views\Project\FeaturedListViewData.cs" />
     <Compile Include="Views\Project\FullProjectListSimple.cs" />
@@ -1376,7 +1376,7 @@
     <Content Include="Views\Shared\ProjectUpdateDiffControls\ProjectFundingRequestsDetail.cshtml" />
     <Content Include="Views\ProjectCreate\ExpectedFunding.cshtml" />
     <Content Include="Views\Shared\EditorTemplates\MoneyWholeNumber.cshtml" />
-    <Content Include="Views\Project\FundingRequestSheet.cshtml" />
+    <Content Include="Views\Project\ForwardLookingFactSheet.cshtml" />
     <Reference Include="Antlr3.Runtime">
       <HintPath>..\NuGetPackages\Antlr.3.5.0.2\lib\Antlr3.Runtime.dll</HintPath>
     </Reference>
@@ -3193,7 +3193,7 @@
     <Content Include="Views\Home\Index.cshtml" />
     <Content Include="Views\ProjectExternalLink\EditProjectExternalLinks.cshtml" />
     <Content Include="Views\Project\EditFeaturedProjects.cshtml" />
-    <Content Include="Views\Project\FactSheet.cshtml" />
+    <Content Include="Views\Project\BackwardLookingFactSheet.cshtml" />
     <Content Include="Views\Project\FeaturedList.cshtml" />
     <Content Include="Views\Project\FullProjectListSimple.cshtml" />
     <Content Include="Views\Project\Index.cshtml" />
