+++ Source/ProjectFirma.Web/Views/Project/BackwardLookingFactSheetViewData.cs
@@ -1,6 +1,6 @@
 ﻿/*-----------------------------------------------------------------------
-<copyright file="FactSheetViewData.cs" company="Tahoe Regional Planning Agency">
-Copyright (c) Tahoe Regional Planning Agency. All rights reserved.
+<copyright file="FactSheetViewData.cs" company="Tahoe Regional Planning Agency and Sitka Technology Group">
+Copyright (c) Tahoe Regional Planning Agency and Sitka Technology Group. All rights reserved.
 <author>Sitka Technology Group</author>
 </copyright>
 
@@ -18,12 +18,11 @@ GNU Affero General Public License <http://www.gnu.org/licenses/> for more detail
 Source code is available upon request via <support@sitkatech.com>.
 </license>
 -----------------------------------------------------------------------*/
-
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
-using LtInfo.Common;
+using ProjectFirma.Web.Controllers;
 using ProjectFirma.Web.Models;
 using ProjectFirma.Web.Views.Map;
 using ProjectFirma.Web.Views.Shared;
@@ -33,52 +32,78 @@ using ProjectFirma.Web.Common;
 
 namespace ProjectFirma.Web.Views.Project
 {
-    public class FundingRequestSheetViewData : ProjectViewData
+    public class BackwardLookingFactSheetViewData : ProjectViewData
     {
+        public readonly ImageGalleryViewData ImageGalleryViewData;
         public readonly ProjectLocationSummaryViewData ProjectLocationSummaryViewData;
-        public readonly List<IGrouping<Models.PerformanceMeasure, PerformanceMeasureExpected>> PerformanceMeasureExpectedValues;
-        public readonly List<GooglePieChartSlice> FundingSourceRequestAmountGooglePieChartSlices;
+        public readonly List<IGrouping<Models.PerformanceMeasure, ProjectPerformanceMeasureReportingPeriodValue>> PerformanceMeasureReportedValues;
+        public readonly List<GooglePieChartSlice> ExpenditureGooglePieChartSlices;
+        public readonly string ChartID;
+        //public readonly List<GooglePieChartSlice> ExpenditureGooglePieChartSlices;
         public readonly Models.ProjectImage KeyPhoto;
         public readonly List<IGrouping<ProjectImageTiming, Models.ProjectImage>> ProjectImagesExceptKeyPhotoGroupedByTiming;
         public readonly int ProjectImagesPerTimingGroup;
+        public readonly List<string> ChartColorRange;
         public readonly List<Models.Classification> Classifications;
         public readonly GoogleChartJson GoogleChartJson;
-        public readonly string SupportingAgenciesForDisplay;
-        public readonly string FundingRequest;
         public readonly int CalculatedChartHeight;
 
         public readonly string TaxonomyColor;
-        public readonly string TaxonomyTierOneDisplayName;
         public readonly string TaxonomyTierOneName;
-        private readonly string TaxonomyTierTwoName;
+        public readonly string TaxonomyTierTwoName;
         public readonly string ClassificationDisplayNamePluralized;
 
-        public FundingRequestSheetViewData(Person currentPerson,
-            Models.Project project,
-            ProjectLocationSummaryMapInitJson projectLocationSummaryMapInitJson,
-            GoogleChartJson googleChartJson,
-            List<GooglePieChartSlice> fundingSourceRequestAmountGooglePieChartSlices) : base(currentPerson, project)
+        public readonly string TaxonomyTierOneDisplayName;
+        public readonly Person PrimaryContactPerson;
+
+        public BackwardLookingFactSheetViewData(Person currentPerson, Models.Project project, ProjectLocationSummaryMapInitJson projectLocationSummaryMapInitJson, GoogleChartJson getProjectFactSheetGoogleChart,
+            List<GooglePieChartSlice> getExpenditureGooglePieChartSlices, List<string> chartColorRange) : base(currentPerson, project)
         {
             PageTitle = project.DisplayName;
             BreadCrumbTitle = "Fact Sheet";
 
-            PerformanceMeasureExpectedValues = project.PerformanceMeasureExpecteds.GroupBy(x => x.PerformanceMeasure, new HavePrimaryKeyComparer<Models.PerformanceMeasure>())
+            const bool userCanAddPhotosToThisProject = false;
+            var newPhotoForProjectUrl = string.Empty;
+            var selectKeyImageUrl = string.Empty;
+            var galleryName = $"ProjectImage{project.ProjectID}";
+            ImageGalleryViewData = new ImageGalleryViewData(currentPerson,
+                galleryName,
+                project.ProjectImages,
+                userCanAddPhotosToThisProject,
+                newPhotoForProjectUrl,
+                selectKeyImageUrl,
+                true,
+                x => x.CaptionOnFullView,
+                "Photo");
+
+            PerformanceMeasureReportedValues = project.PerformanceMeasureActuals.Select(x => x.PerformanceMeasure)
+                .Distinct(new HavePrimaryKeyComparer<Models.PerformanceMeasure>())
+                .SelectMany(x =>
+                    x.GetProjectPerformanceMeasureSubcategoryOptionReportedValues(x, new List<int> {project.ProjectID}))
+                .OrderByDescending(pma => pma.CalendarYear)
+                .ThenBy(pma => pma.PerformanceMeasureReportingPeriod.PerformanceMeasure.PerformanceMeasureID)
+                .GroupBy(x => x.PerformanceMeasureReportingPeriod.PerformanceMeasure,
+                    new HavePrimaryKeyComparer<Models.PerformanceMeasure>())
                 .OrderBy(x => x.Key.PerformanceMeasureDisplayName).ToList();
             ProjectLocationSummaryViewData = new ProjectLocationSummaryViewData(project, projectLocationSummaryMapInitJson);
 
+            ChartID = $"fundingChartForProject{project.ProjectID}";
+            //FundingSourceExpenditures = project.GetExpenditureGooglePieChartSlices();
             KeyPhoto = project.KeyPhoto;
-            ProjectImagesExceptKeyPhotoGroupedByTiming = project.ProjectImages.Where(x => !x.IsKeyPhoto && x.ProjectImageTiming != ProjectImageTiming.Unknown && !x.ExcludeFromFactSheet)
-                .GroupBy(x => x.ProjectImageTiming).OrderBy(x => x.Key.SortOrder).ToList();
+            ProjectImagesExceptKeyPhotoGroupedByTiming =
+                project.ProjectImages.Where(x => !x.IsKeyPhoto && x.ProjectImageTiming != ProjectImageTiming.Unknown && !x.ExcludeFromFactSheet)
+                    .GroupBy(x => x.ProjectImageTiming)
+                    .OrderBy(x => x.Key.SortOrder)
+                    .ToList();
             ProjectImagesPerTimingGroup = ProjectImagesExceptKeyPhotoGroupedByTiming.Count == 1 ? 6 : 2;
             Classifications = project.ProjectClassifications.Select(x => x.Classification).OrderBy(x => x.DisplayName).ToList();
 
-            GoogleChartJson = googleChartJson;
-            FundingSourceRequestAmountGooglePieChartSlices = fundingSourceRequestAmountGooglePieChartSlices;
+            GoogleChartJson = getProjectFactSheetGoogleChart;
 
+            ExpenditureGooglePieChartSlices = getExpenditureGooglePieChartSlices;
+            ChartColorRange = chartColorRange;
             //Dynamically resize chart based on how much space the legend requires
-            CalculatedChartHeight = 477 - (FundingSourceRequestAmountGooglePieChartSlices.Count <= 2
-                                        ? FundingSourceRequestAmountGooglePieChartSlices.Count * 24
-                                        : FundingSourceRequestAmountGooglePieChartSlices.Count * 20);
+            CalculatedChartHeight = 435 - ExpenditureGooglePieChartSlices.Count * 19;
 
             if (project.TaxonomyTierOne == null)
             {
@@ -103,32 +128,24 @@ namespace ProjectFirma.Web.Views.Project
                             $"ProjectFirma currently only supports up to a 3-tier taxonomy; number of taxonomy tiers is {MultiTenantHelpers.GetNumberOfTaxonomyTiers()}");
                 }
             }
-
             TaxonomyTierOneName = project.TaxonomyTierOne == null ? $"{Models.FieldDefinition.Project.GetFieldDefinitionLabel()} Taxonomy Not Set" : project.TaxonomyTierOne.DisplayName;
             TaxonomyTierTwoName = project.TaxonomyTierOne == null ? $"{Models.FieldDefinition.Project.GetFieldDefinitionLabel()} Taxonomy Not Set" : project.TaxonomyTierOne.TaxonomyTierTwo.DisplayName;
             TaxonomyTierOneDisplayName = Models.FieldDefinition.TaxonomyTierOne.GetFieldDefinitionLabel();
             ClassificationDisplayNamePluralized = Models.FieldDefinition.Classification.GetFieldDefinitionLabelPluralized();
-
-            SupportingAgenciesForDisplay = project.ProjectFundingSourceRequests.Any()
-                ? string.Join(", ", project.ProjectFundingSourceRequests.Select(x => x.FundingSource.Organization.DisplayName).OrderBy(x => x))
-                : "?";
-            FundingRequest = project.ProjectFundingSourceRequests.Any() ? project.ProjectFundingSourceRequests.Sum(x => x.UnsecuredAmount).ToStringCurrency() : "?";
+            PrimaryContactPerson = project.GetPrimaryContact();
         }
 
-        public HtmlString LegendHtml
+        public HtmlString GetLegendHTML()
         {
-            get
+            var legendHtml = "<div>";
+            foreach (var googlePieChartSlice in ExpenditureGooglePieChartSlices.OrderBy(x => x.SortOrder))
             {
-                var legendHtml = "<div>";
-                foreach (var googlePieChartSlice in FundingSourceRequestAmountGooglePieChartSlices.OrderBy(x => x.SortOrder))
-                {
-                    legendHtml += "<div class='chartLegendColorBox' style='display:inline-block; border: solid 6px " + googlePieChartSlice.Color + "'></div> ";
-                    legendHtml += "<div style='display:inline-block' >" + googlePieChartSlice.Label + "</div>";
-                    legendHtml += "<br>";
-                }
-                legendHtml += "</div>";
-                return new HtmlString(legendHtml);
+                legendHtml += "<div class='chartLegendColorBox' style='display:inline-block; border: solid 6px " + googlePieChartSlice.Color + "'></div> ";
+                legendHtml += "<div style='display:inline-block' >" + googlePieChartSlice.Label + ": " + googlePieChartSlice.Value.ToString("C0") + "</div>";
+                legendHtml += "<br>";
             }
+            legendHtml += "</div>";
+            return new HtmlString(legendHtml);
         }
     }
 }
